lmaoai-backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
docx==0.2.4
email-validator==1.3.1
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.2
orjson==3.8.9
packaging==23.0
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.5.0
platformdirs==3.2.0
pycparser==2.21
pydantic==1.10.7
pypdf==3.7.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
Wand==0.6.11
watchfiles==0.19.0
websockets==11.0
yarl==1.8.2
